Fix CI link failure: move DE-9IM filter validation into SpatialQuery
The regex-based validation for DE-9IM filter strings was compiled into
the `engine` library, but its only caller, `SpatialQuery.cpp`, lives in
`sparqlParser`, which `engine` depends on (not the reverse). This broke
linking of qlever-index/qlever-server. Since SpatialQuery is the only
caller, fold the validation functions directly into SpatialQuery.{h,cpp}
instead of introducing a separate file.
