+++ src/engine/SpatialJoinConfig.h
@@ -13,7 +13,6 @@
 #include <array>
 #include <cstddef>
 #include <optional>
-#include <string_view>
 #include <variant>
 
 #include "engine/SpatialJoinType.h"
@@ -29,37 +28,7 @@
 // https://en.wikipedia.org/wiki/DE-9IM). Stored without a null terminator.
 using De9imFilterString = std::array<char, 9>;
 
-// If `filter` is a syntactically valid DE-9IM filter pattern (i.e. exactly 9
-// characters, each one of `0`-`2`, `T`/`t`, `F`/`f`, or `*`, see
-// `De9imFilterString` above), return it as a `De9imFilterString`, else
-// `std::nullopt`. Note: this does not check whether the pattern can match
-// disjoint geometries, see `de9imFilterCanMatchDisjoint` below for that.
-std::optional<De9imFilterString> parseDe9imFilterString(
-    std::string_view filter);
-
-// Whether the given (syntactically valid) DE-9IM `filter` could match a
-// disjoint pair of geometries. Patterns for which this holds (e.g.
-// `*********` or the literal disjoint pattern `FF*FF****`) are unsupported:
-// the pinned `libspatialjoin` never enumerates disjoint candidate pairs to
-// its callback (see `Sweeper::doDE9IMCheck`), regardless of the configured
-// filter, so accepting such a pattern would silently omit matching disjoint
-// pairs from the result.
-//
-// The DE-9IM matrix entries are ordered II, IB, IE, BI, BB, BE, EI, EB, EE. A
-// pair of geometries is disjoint iff II, IB, BI, and BB (indices 0, 1, 3, 4)
-// are all `F`. A filter character only excludes `F` if it is a digit, `T`, or
-// `t`; `*` and `F`/`f` both admit it. If all four of these positions admit
-// `F`, the pattern could match a disjoint pair.
-bool de9imFilterCanMatchDisjoint(const De9imFilterString& filter);
-
-// If `filter` is a syntactically valid DE-9IM filter pattern that cannot
-// match a disjoint pair of geometries, return it as a `De9imFilterString`,
-// else `std::nullopt`. See `parseDe9imFilterString` and
-// `de9imFilterCanMatchDisjoint` above, which this combines and which should
-// be used directly if the two failure cases need to be reported separately
-// (as in `SpatialQuery.cpp`).
-std::optional<De9imFilterString> validateDe9imFilterString(
-    std::string_view filter);
+// Parsing and validation of `De9imFilterString`s, see `parser/SpatialQuery.h`.
 
 // A nearest neighbor search with optionally a maximum distance.
 struct NearestNeighborsConfig {
