+++ src/parser/SpatialQuery.cpp
@@ -5,11 +5,45 @@
 
 #include "parser/SpatialQuery.h"
 
+#include <algorithm>
+
 #include "engine/SpatialJoinConfig.h"
 #include "parser/MagicServiceIriConstants.h"
 #include "parser/NormalizedString.h"
 #include "parser/PayloadVariables.h"
 #include "parser/SparqlTriple.h"
+
+namespace {
+constexpr ctll::fixed_string de9imFilterRegex = "[0-2TtFf*]{9}";
+}  // namespace
+
+// ____________________________________________________________________________
+std::optional<De9imFilterString> parseDe9imFilterString(
+    std::string_view filter) {
+  if (!ctre::match<de9imFilterRegex>(filter)) {
+    return std::nullopt;
+  }
+  De9imFilterString result{};
+  std::ranges::copy(filter, result.begin());
+  return result;
+}
+
+// ____________________________________________________________________________
+bool de9imFilterCanMatchDisjoint(const De9imFilterString& filter) {
+  auto admitsF = [](char c) { return c == '*' || c == 'F' || c == 'f'; };
+  return admitsF(filter[0]) && admitsF(filter[1]) && admitsF(filter[3]) &&
+         admitsF(filter[4]);
+}
+
+// ____________________________________________________________________________
+std::optional<De9imFilterString> validateDe9imFilterString(
+    std::string_view filter) {
+  auto result = parseDe9imFilterString(filter);
+  if (!result.has_value() || de9imFilterCanMatchDisjoint(result.value())) {
+    return std::nullopt;
+  }
+  return result;
+}
 
 namespace parsedQuery {
